refactor: quorum members: remove QuorumObserverParent; participant is no more child of 'observer'
